macos: move ScreenText to OSSurfaceView
This allows it to be shared with a future iOS implementation.
